+++ src/common/styles.css
@@ -1,0 +1,38 @@
+:root {
+  --primary-color: hsl(204, 6%, 17%);
+  --accent-color: hsl(2, 56%, 54%);
+}
+html {
+  font-size: 62.5%;
+}
+body {
+  font-family: 'Roboto';
+  margin: 0 auto;
+}
+h1 {
+  font-size: 4.2rem;
+  font-weight: 700;
+}
+h2 {
+  font-size: 3.6rem;
+  font-weight: 700;
+}
+h3 {
+  font-size: 2.4rem;
+  font-weight: 700;
+}
+h4,
+h5,
+h6 {
+  font-size: 1.8rem;
+  font-weight: 400;
+}
+p {
+  font-size: 1.6rem;
+  font-weight: 300;
+}
+
+main {
+  max-width: 1000px;
+  margin: auto;
+}
