+++ src/features/editor/MetadataBlock.tsx
@@ -1,5 +1,6 @@
 import { MetadataModel, MetadataStatus } from '@/common/types';
 import { Dispatch, SetStateAction } from 'react';
+import styles from './Editor.module.css';
 
 export function MetadataBlock({
   metadata,
@@ -10,29 +11,14 @@ export function MetadataBlock({
 }) {
   const metadataClone: typeof metadata = JSON.parse(JSON.stringify(metadata));
   return (
-    <div className="MetadataBlock">
+    <>
       <input
-        className="MetadataTitle"
+        className={styles.title}
         onChange={(e) =>
           setMetadata({ ...metadataClone, title: e.target.value })
         }
         defaultValue={metadata.title}
       />
-      <select
-        className="MetadataStatus"
-        onChange={(e) => {
-          let input = e.target.value;
-          let isValid = (input: string): input is MetadataStatus =>
-            ['draft', 'outline', 'live'].includes(input);
-          let status = isValid(input) ? input : metadata.status;
-          return setMetadata({ ...metadataClone, status });
-        }}
-        defaultValue={metadata.status}
-      >
-        <option value="outline">outline</option>
-        <option value="draft">draft</option>
-        <option value="live">live</option>
-      </select>
-    </div>
+    </>
   );
 }
