+++ src/pages/_app.tsx
@@ -3,12 +3,17 @@ import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
 import { ReactQueryDevtools } from '@tanstack/react-query-devtools';
 import { useState } from 'react';
 import '@common/styles.css';
+import { Roboto } from '@next/font/google';
+
+const roboto = Roboto({ subsets: ['latin'], weight: ['300', '400', '700'] });
 
 export default function MyApp({ Component, pageProps }: AppProps) {
   const [queryClient] = useState(() => new QueryClient());
   return (
     <QueryClientProvider client={queryClient}>
-      <Component {...pageProps} />
+      <main className={roboto.className}>
+        <Component {...pageProps} />
+      </main>
       <ReactQueryDevtools initialIsOpen={false} />
     </QueryClientProvider>
   );
